requirements.txt
pandas === 1.1.3
numpy === 1.18.5
Pillow === 8.0.1
matplotlib === 3.3.2
opencv-python === 4.4.0.46
seaborn === 0.11.0
scikit-learn === 0.23.2
imageio === 2.9.0
scikit-image === 0.17.2
tensorflow === 2.3.1
Keras === 2.3.1
PyYAML === 5.3.1